+++ app/src/main/java/com/example/dopaminemoa/presentation/myvideo/MyVideoFragment.kt
@@ -1,4 +1,4 @@
-package com.example.dopaminemoa.presentation.myvideos
+package com.example.dopaminemoa.presentation.myvideo
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.dopaminemoa.R
 
-class MyVideosFragment : Fragment() {
+class MyVideoFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_my_videos, container, false)
+        return inflater.inflate(R.layout.fragment_my_video, container, false)
     }
 
 }
