+++ app/src/main/res/layout/fragment_my_video.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.myvideos.MyVideosFragment">
+    tools:context=".presentation.myvideo.MyVideoFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
